+++ src/Arguments/KeyVault/Key/KeyListArgument.cs
@@ -1,0 +1,7 @@
+﻿namespace AzureMcp.Arguments.KeyVault.Key
+{
+    public class KeyListArgument : BaseKeyVaultArguments
+    {
+        public bool IncludeManagedKeys { get; set; } = false;
+    }
+}