+++ src/Models/Argument/ArgumentDefinitions.cs
@@ -416,6 +416,7 @@ public static class ArgumentDefinitions
         public const string VaultNameParam = "vault";
         public const string KeyNameParam = "key";
         public const string KeyTypeParam = "key-type";
+        public const string IncludeManagedKeysParam = "include-managed";
 
         public static readonly ArgumentDefinition<string> VaultName = new(
             VaultNameParam,
@@ -431,6 +432,11 @@ public static class ArgumentDefinitions
             KeyTypeParam,
             "The type of key to create (RSA, EC).",
             required: true);
+
+        public static readonly ArgumentDefinition<bool> IncludeManagedKeys = new(
+            IncludeManagedKeysParam,
+            "Whether or not to include managed keys in results.",
+            required: false);
     }
 
     public static class ServiceBus